Fix erroneous CP-1252 character in response body
It would appear that the BLDSS API is returning at least the '
incorrectly encoded in CP-1252 in the UTF-8 response. This is almost
certainly a problem in the underlying data.

This is therefore a hack that just replaces it with the correct
character until such a time as the BL fix it themselves. I emailed them
on 16 Oct 2018.

It may be necessary to add additional characters to the replacement
mapping table, this is in the _clean_response sub.