Fix request/header escaping mismatch
Authorised requests to the BLDSS are exceedingly picky and we've had a
bug in how we make them since day one, but it only came to light with
the occurrence of a specific character ":" in item metadata.

The authorised requests comprise the request URL, possibly a body and
the the authorisation header. In GET requests the item metadata must be
identical to that used to generate the authorisation token in the
header. In this case it was not. It seems that URI::query_form and
URI::Escape do things a bit differently, so we were getting a disparity,
which was causing an incorrect authorisation token to be generated.