Add deprecation warnings to unused subs
There are a number of subs that exist but have never been used. These
add a lot of complication to trying to understand the code. As such,
I've added deprecation warnings to them in order to a) flag them as
unused b) Alert us if they are used in some way

I'm not sure they should be removed as some of them maybe useful in the
future. However, that said, they don't really reflect the state of the
code as it currently stands, so their usefulness is questionable.