Make search results datatable and add column class
The search results table should be a datatable.

We need the ability to be able to hide columns in the search results.
Since adding Koha's columns settings would require changes to core,
that we'd never get through community since it's backend dependent, I've
added data generated classes to the table columns, so they can be hidden
with CSS.