Switch to loan request when appropriate
As per issue #12.

Note: This only affects unmediated requests.

If a site is making use of the <loan_book_if_available> flag (which
can now appear in a <branch>, <borrower_category> or the main
<interlibrary_loans> block) and the user requests an entire book, i.e. a
book type with no additional submitted metadata, we automatically request the
"loan" format. Previously we were using whatever format was defined in
the <default_formats> block, which could lead to requests for entire
items being made as downloads, which would lead to the infamous "Invalid
request: at least one Item of Interest level piece of metadata is
required" error.